requirements.txt
boto3==1.34.151
asgiref==3.7.2
celery==5.4.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.0.1
crispy-bootstrap5==2023.10
cryptography==42.0.4
defusedxml==0.7.1
Django==5.0.3
django-ace==1.32.4
django-allauth==0.61.1
django-admin-autocomplete-filter==0.7.1
django-admin-interface==0.28.6
django-apptemplates==1.5
django-clearcache==1.2.1
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-debug-toolbar==4.2.0
django-environ==0.11.2
django-filter==24.2
django-ninja==1.2.2
dj-database-url==2.2.0
environ==1.0
gunicorn==22.0.0
idna==3.7
oauthlib==3.2.2
packaging==23.1
psycopg==3.1.18
psycopg-binary==3.1.18
pycparser==2.21
PyJWT==2.6.0
python3-openid==3.2.0
orjson==3.10.6
redis==5.0.4
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==2.0.7
whitenoise==6.6.0
